moving publishing logic out of workbook methods (in case we want to support psd, prep flow, etc)
